requirements.txt
psycopg2-binary
python-decouple
flask_bootstrap
flask_wtf
dynaconf
